Add Exa as ls replacement.
Add new folder for programming stuff and put in an install script.
Add Rust as language to be installed including cargo.
Adjust the Zsh runtime configuration to extend the PATH for Rust and
source cargo.
Extend the shell install script by installing Exa.
Add a new alias to use Exa instead of ls.