Rename shared resource to shared directories
The resources are not shared, but the directories. This gets
strengthened now by using the same directory name within the role files
folder, as for the actual shared directory where to link in. This
allowed to simply the loop of the shared directory task block.

Furthermore the autostart directory has been removed, since this should
not been used anymore.