Use custom enlarged Varant base box
The original box is limited to 2GB of space. It is possible to increase
the size of the virtual disk. But is does not work to resize the root
partition while the machine is busy. This requires some external
bootable system to do so. Since this can't be automated via the
provision script and therefore breaks the automated test pipe, a custom
Vagrant base box has been created, having a virtual disk size of 20GB.
The box is available on VagrantCloud, too large to store in the
repository.