Export Vagrant provision script
The script has become quite long. So far the argument to keep it inside
the Vagrantfile was that it would pollute the repository structure as
a single script. Since it there are more script about to come, it makes
sense to collect them in their own directory. Therefore also the
provision script can be exported, making the Vagrantfile more easy to
read.