Remove usage of sub-shell in i3 configuration
This was (is) only necessary for the usage of scripts that are
"published" as alias. Since they do now get installed to /usr/bin, this
is not of an issue anymore.
This also removed the obsolete music mode. This should not be part of
the i3 configuration at all. It will become migrated at a later point
as hotkeys for the Mopidy role.