Switch from Xbindkeys to Sxhkd
Sxhkd has two main advantages. First it allows the usage of combos. This
means not all keys must be pressed together, but can be orchestrated
after each other. This is especially useful when sub-modes are (still)
missing. This is not yet used, but will be soon.
Moreover does it support natively to have multiple files defining the
hotkeys. It even allows to provide a directory with a glob.
Finally, just a tiny positive point it that it provides a filetype with
highlighting for Vim.