Add Polybar as Powerstatus10k replacement
This is not feature complete yet. Missing segments are mail, calendar
and bluetooth. All are not part of the in-built segments and must be
added by hand. They probably get imported/inspired by the according
Powerstatus10k segments.
The in-built segments partially look not so great. But custom segments
do not have the full capabilities. Stick to them for the moment.
Figure out if it is feasible to let other roles adding segments. The
current configuration uses already an import of a whole directory. But
the order in the segment list and the separator modules is tricky.