Change base role handling for systemd units
So far it was only possible to define scope for all units. There was no
chance to differ them. This is a problem when a role wants to establish
units for the user and system scope.
Therefore, the single task to establish the units has been split into
three. The first acts on a new list variable called
`systemd_units_user`. All units in this list are enabled for the user
scope. The same for the second with the `systemd_units_system` variable.
Finally the last task is only for the custom unit files. Here the
variable `systemd_units_custom_scope` decides about their scope.
The caveat is that different custom unit files can't differ in their
scope. This is not a requirement yet and pretty unlikely to happen.
Therefore I stick to this approach for now.