Replace Ansible setup with NixOS template
NixOS will become the third generation of this repository with a complete
re-work of the full setup. Thought this time is is quite different. I still love
the old Ansible setup and will definitely tag it. I'll also just merge this to
master if it 100% surpassed it and is fully usable. Looking forward to see great
changes.