Delete editorconfig file as it is redundant
In my opinion editorconfigs are obsolete. As long as a repository enforces
a linter and autoformatting tool usage, it doesn't matter what a developers
local editor does. Code can be formatted locally on pulls and push hooks. How
the code looks locally in comparison to how it looks on the remote doesn't
matter.