Initiate set of hosts
Though it is kinda useless to create all these new hosts, it is just a first
getting in touch with the repository. And it is my intention to use variating
setups for all these hosts. Each host relates to a physical machine I use.

Fun fact:
The host names are all related to horse breeds. The "attributes" I associate
with the machines of the host get mapped to a horse breed which can be
associated with them.
  Trakehner - breed of my own horse her father, sportive, strong war horses
            -> computer for private/personal stuff
  Camargue  - breed known for its strong and though work horses
            -> computer for work with more hardware
  Shetland  - small ponies who like to stink
            -> old former Windows computer for experiments