Add my own user instead of generic one
This is just a very basic user for the moment. Just for the sake of having it
personalized. This will become kinda obsolete in future anyway. My plans are to
have an independent user on an external drive with systemd-homed. But until
then, I'll first work in the system setups. So I need/want to have a user.