Rofi: fix theme and keymaps for latest version
Since the latest version the theme was no more working properly as it
wasn't centered on screen. Also some elements haven't inherited the
colors.
Furthermore the default key mappings have changed and lead to errors.
A key combo can't be re-used. The default mapping using it must be
overwritten to make the combo available again.