Picom: remove configuration file
For some reason, with the latest update of Picom, it was slowing down
the desktop environment immensely and I was in need to disable Picom.
Figured out that the configuration was causing the issue. As I don't
really need this, I simply removed it. Too lazy to figure out what is
the underlying problem.